+++ MeasurableSettings.php
@@ -21,7 +21,19 @@ use Piwik\Validators\Exception as ValidatorException;
 class MeasurableSettings extends BaseMeasurableSettings
 {
     /** @var Setting */
+    public $isEnabled;
+
+    /** @var Setting */
+    public $hasExtendedTimeout;
+
+    /** @var Setting */
+    public $runCount;
+
+    /** @var Setting */
     public $emulatedDevice;
+
+    /** @var Setting */
+    public $hasGroupedUrls;
 
     /** @var Setting */
     public $hasExtraHttpHeader;
@@ -32,9 +44,6 @@ class MeasurableSettings extends BaseMeasurableSettings
     /** @var Setting */
     public $extraHttpHeaderValue;
 
-    /** @var Setting */
-    public $runCount;
-
     /**
      * Initialise plugin settings.
      *
@@ -43,13 +52,47 @@ class MeasurableSettings extends BaseMeasurableSettings
      */
     protected function init()
     {
-        Piwik::checkUserHasSomeAdminAccess();
+        Piwik::checkUserHasSomeViewAccess();
 
+        $this->isEnabled = $this->makeIsEnabledSetting();
+        $this->hasExtendedTimeout = $this->makeHasExtendedTimeoutSetting();
         $this->runCount = $this->makeRunCountSetting();
         $this->emulatedDevice = $this->makeEmulatedDeviceSetting();
+        $this->hasGroupedUrls = $this->makeHasGroupedUrlsSetting();
         $this->hasExtraHttpHeader = $this->makeHasExtraHttpHeaderSetting();
         $this->extraHttpHeaderKey = $this->makeExtraHttpHeaderKeySetting();
         $this->extraHttpHeaderValue = $this->makeExtraHttpHeaderValueSetting();
+    }
+
+    /**
+     * Create is enabled setting.
+     *
+     * @return MeasurableSetting
+     * @throws ValidatorException|Exception
+     */
+    private function makeIsEnabledSetting()
+    {
+        return $this->makeSetting('is_enabled', true, FieldConfig::TYPE_BOOL, function (FieldConfig $field) {
+            $field->title = Piwik::translate('PerformanceAudit_Settings_IsEnabled_Title');
+            $field->inlineHelp = Piwik::translate('PerformanceAudit_Settings_IsEnabled_Help');
+            $field->uiControl = FieldConfig::UI_CONTROL_CHECKBOX;
+        });
+    }
+
+    /**
+     * Create has extended timeout setting.
+     *
+     * @return MeasurableSetting
+     * @throws ValidatorException|Exception
+     */
+    private function makeHasExtendedTimeoutSetting()
+    {
+        return $this->makeSetting('has_extended_timeout', false, FieldConfig::TYPE_BOOL, function (FieldConfig $field) {
+            $field->title = Piwik::translate('PerformanceAudit_Settings_HasExtendedTimeout_Title');
+            $field->inlineHelp = Piwik::translate('PerformanceAudit_Settings_HasExtendedTimeout_Help');
+            $field->condition = 'is_enabled';
+            $field->uiControl = FieldConfig::UI_CONTROL_CHECKBOX;
+        });
     }
 
     /**
@@ -63,6 +106,7 @@ class MeasurableSettings extends BaseMeasurableSettings
         return $this->makeSetting('run_count', 3, FieldConfig::TYPE_INT, function (FieldConfig $field) {
             $field->title = Piwik::translate('PerformanceAudit_Settings_RunCount_Title');
             $field->inlineHelp = Piwik::translate('PerformanceAudit_Settings_RunCount_Help');
+            $field->condition = 'is_enabled';
             $field->uiControl = FieldConfig::UI_CONTROL_TEXT;
             $field->validate = function ($value) {
                 if (empty($value) && $value != 0) {
@@ -86,6 +130,7 @@ class MeasurableSettings extends BaseMeasurableSettings
         return $this->makeSetting('emulated_device', EmulatedDevice::__default, FieldConfig::TYPE_STRING, function (FieldConfig $field) {
             $field->title = Piwik::translate('PerformanceAudit_Settings_EmulatedDevice_Title');
             $field->inlineHelp = Piwik::translate('PerformanceAudit_Settings_EmulatedDevice_Help');
+            $field->condition = 'is_enabled';
             $field->uiControl = FieldConfig::UI_CONTROL_SINGLE_SELECT;
             $field->availableValues = [
                 EmulatedDevice::Desktop => ucfirst(Piwik::translate('PerformanceAudit_EnvironmentDesktop')),
@@ -104,6 +149,22 @@ class MeasurableSettings extends BaseMeasurableSettings
     }
 
     /**
+     * Create a grouped URLs setting.
+     *
+     * @return MeasurableSetting
+     * @throws ValidatorException|Exception
+     */
+    private function makeHasGroupedUrlsSetting()
+    {
+        return $this->makeSetting('has_grouped_urls', false, FieldConfig::TYPE_BOOL, function (FieldConfig $field) {
+            $field->title = Piwik::translate('PerformanceAudit_Settings_HasGroupedUrls_Title');
+            $field->inlineHelp = Piwik::translate('PerformanceAudit_Settings_HasGroupedUrls_Help');
+            $field->condition = 'is_enabled';
+            $field->uiControl = FieldConfig::UI_CONTROL_CHECKBOX;
+        });
+    }
+
+    /**
      * Create has extra HTTP header setting.
      *
      * @return MeasurableSetting
@@ -114,6 +175,7 @@ class MeasurableSettings extends BaseMeasurableSettings
         return $this->makeSetting('has_extra_http_header', false, FieldConfig::TYPE_BOOL, function (FieldConfig $field) {
             $field->title = Piwik::translate('PerformanceAudit_Settings_HasExtraHttpHeader_Title');
             $field->inlineHelp = Piwik::translate('PerformanceAudit_Settings_HasExtraHttpHeader_Help');
+            $field->condition = 'is_enabled';
             $field->uiControl = FieldConfig::UI_CONTROL_CHECKBOX;
         });
     }
@@ -131,14 +193,14 @@ class MeasurableSettings extends BaseMeasurableSettings
         return $this->makeSetting('extra_http_header_key', '', FieldConfig::TYPE_STRING, function (FieldConfig $field) use ($self) {
             $field->title = Piwik::translate('PerformanceAudit_Settings_ExtraHttpHeaderKey_Title');
             $field->inlineHelp = Piwik::translate('PerformanceAudit_Settings_ExtraHttpHeaderKey_Help');
-            $field->condition = 'has_extra_http_header';
+            $field->condition = 'is_enabled && has_extra_http_header';
             $field->uiControl = FieldConfig::UI_CONTROL_SINGLE_SELECT;
             $field->availableValues = [
                 'Authorization' => Piwik::translate('PerformanceAudit_Settings_ExtraHttpHeaderKey_Authorization'),
                 'Cookie' => Piwik::translate('PerformanceAudit_Settings_ExtraHttpHeaderKey_Cookie'),
             ];
             $field->validate = function ($value) use ($self, $field) {
-                if ($self->getSetting($field->condition)->getValue()) {
+                if ($self->hasExtraHttpHeader()) {
                     if (empty($value)) {
                         throw new ValidatorException(Piwik::translate('General_ValidatorErrorEmptyValue'));
                     }
@@ -163,10 +225,10 @@ class MeasurableSettings extends BaseMeasurableSettings
         return $this->makeSetting('extra_http_header_value', '', FieldConfig::TYPE_STRING, function (FieldConfig $field) use ($self) {
             $field->title = Piwik::translate('PerformanceAudit_Settings_ExtraHttpHeaderValue_Title');
             $field->inlineHelp = Piwik::translate('PerformanceAudit_Settings_ExtraHttpHeaderValue_Help');
-            $field->condition = 'has_extra_http_header';
+            $field->condition = 'is_enabled && has_extra_http_header';
             $field->uiControl = FieldConfig::UI_CONTROL_TEXT;
             $field->validate = function ($value) use ($self, $field) {
-                if ($self->getSetting($field->condition)->getValue()) {
+                if ($self->hasExtraHttpHeader()) {
                     if (empty($value)) {
                         throw new ValidatorException(Piwik::translate('General_ValidatorErrorEmptyValue'));
                     }
@@ -176,5 +238,65 @@ class MeasurableSettings extends BaseMeasurableSettings
                 }
             };
         });
+    }
+
+    /**
+     * Returns array of runs of site.
+     *
+     * @return array
+     */
+    public function getRuns()
+    {
+        return range(1, (int) $this->getSetting('run_count')->getValue());
+    }
+
+    /**
+     * Returns list of emulated devices of site.
+     *
+     * @return array
+     */
+    public function getEmulatedDevicesList()
+    {
+        return EmulatedDevice::getList($this->getSetting('emulated_device')->getValue());
+    }
+
+    /**
+     * Returns if site has audits enabled or not.
+     *
+     * @return bool
+     */
+    public function isAuditEnabled()
+    {
+        return $this->getSetting('is_enabled')->getValue();
+    }
+
+    /**
+     * Returns if site has extended timeout for site audit.
+     *
+     * @return bool
+     */
+    public function hasExtendedTimeout()
+    {
+        return $this->getSetting('has_extended_timeout')->getValue();
+    }
+
+    /**
+     * Returns if site has grouped URLs for site.
+     *
+     * @return bool
+     */
+    public function hasGroupedUrls()
+    {
+        return $this->getSetting('has_grouped_urls')->getValue();
+    }
+
+    /**
+     * Returns if site has extra HTTP header for site.
+     *
+     * @return bool
+     */
+    public function hasExtraHttpHeader()
+    {
+        return $this->getSetting('has_extra_http_header')->getValue();
     }
 }